Move deprecated tests to DeprecatedWholeStageCodegenSuite
